지형 노말맵 로딩, 그림자 다듬기
[+] class GridMesh에 노말맵을 로딩하는 함수 추가
└ 평면인 바닥에 노말맵만 적용시켜서 조명으로 입체감을 표현할 예정
[/] 그림자를 만드는 조명의 위치 적절히 수정
└ 추후 맵을 바꿀 예정이므로 나중에 또 수정해줘야함

+++ 소스/Client/object.cpp
@@ -76,7 +76,7 @@ void GameObject::Update(FLOAT deltaTime)
 				m_textureInfo->frame = 0;
 			else
 			{
-				m_textureInfo->frame = m_texture->GetTextureCount() - 1;
+				m_textureInfo->frame = static_cast<int>(m_texture->GetTextureCount() - 1);
 				m_isDeleted = true;
 			}
 		}
